* commit 'd1ba08e7799393d9c213d479ad99ba681c14e0bc':
  amdgpu 06bc5711a1e78d6f7f129b67d0add2110ec2640f on v5.15.148

+++ drivers/gpu/drm/amd/pm/inc/amdgpu_pm.h
@@ -24,9 +24,8 @@
 #ifndef __AMDGPU_PM_H__
 #define __AMDGPU_PM_H__
 
-struct cg_flag_name
-{
-	u32 flag;
+struct cg_flag_name {
+	u64 flag;
 	const char *name;
 };
